# api/swagger/swagger.yaml
swagger: '2.0'
info:
  description: Welcome to the USA-NPN geospatial services. More description to come!
  version: 0.0.1
  title: National Phenology Network Geospatial Services
  termsOfService: 'TODO: Add terms of service here'
  contact:
    name: The National Phenology Network
    email: support@usanpn.org
    url: www.usanpn.org
host: 'data-dev.usanpn.org:3006'
basePath: /v0
...
schemes:
  - http
  - https
paths:
...
  /si-x/area/boundaryNames:
        x-swagger-router-controller: boundaries
        get:
          summary: get list of available boundary names for FWS or States
          description: Gets list of available boundary names for FWS or States
          tags:
            - boundaries
          operationId: boundaryNames
          parameters:
                  - name: boundaryType
                    in: query
                    description: list the FWS boundaries or the State Boundaries
                    type: string
                    enum:
                      - FWS
                      - States
          responses:
            '200':
              description: successful operation
              schema:
                $ref: '#/definitions/BoundaryNamesResponse'
            '500':
              description: internal server error
              schema:
                $ref: '#/definitions/ErrorResponse'
...
definitions:
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  BoundaryNamesResponse:
      required:
        - boundaryName
      properties:
        boundaryName:
          type: array
          items:
            type: string
...